Add syntax highlight for Rust code between backticks
Rust is a general-purpose, multi-paradigm, compiled programming language
sponsored by Mozilla Research. It is designed to be a "safe, concurrent,
practical language", supporting functional, and imperative-procedural
paradigms. The language grew out of a personal project by Mozilla
employee Graydon Hoare. Mozilla began sponsoring the project in 2009 and
announced it in 2010. The same year, work shifted from the initial
compiler (written in OCaml) to the self-hosting compiler written in
Rust. Known as rustc, it successfully compiled itself in 2011. rustc
uses LLVM as its back end.