Add syntax highlight for Go code between backticks
Go (often referred to as golang) is a Free and open source programming
language created at Google in 2007 by Robert Griesemer, Rob Pike, and
Ken Thompson. It is a compiled, statically typed language in the
tradition of Algol and C, with garbage collection, limited structural
typing, memory safety features and CSP-style concurrent programming
features added. The language was announced in November 2009; it is used
in some of Google's production systems, as well as by other firms.